Moved Monitor to its own file
