+++ src/main/scala/de/jannikarndt/datamover/monitor/Monitoring.scala
@@ -1,0 +1,5 @@
+package de.jannikarndt.datamover.monitor
+
+trait Monitoring {
+    protected val monitor: Monitor = Monitor.getMonitor(getClass.getName)
+}
